Allow `init` to be called more than once
Calling `init` a second time currently results in an error.
This prevents one from re-running the tests after fixing a bug.
This makes it safe it call `init` again; it does't do anything other
than warn the user that the library is already intialized.

Even better might be to make it unnecessary to call `init` at all.
This would be possible from inside an `__init__` function,
but that would prevent you from passing kwargs to `init`.